feat: Remove unused code and update file structure for transaction category management
