+++ nest-cli.json
@@ -2,7 +2,6 @@
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
   "sourceRoot": "src",
-  "entryFile": "infra/http/nest/main",
   "compilerOptions": {
     "plugins": [
       {
