+++ src/common/exception/domain.exception.ts
