+++ src/config/docs/generate-spec.ts
