+++ src/database/exception/client.exception.ts
