--- src/index.spec.ts
@@ -1,7 +1,0 @@
-import { sum } from '.';
-
-describe('index', () => {
-  it('should return 1', () => {
-    expect(sum(1, 0)).toBe(1);
-  });
-});
