--- src/index.ts
@@ -1,3 +1,0 @@
-export function sum (a: number, b: number) {
-  return a + b;
-}
