--- src/infra/http/nest/modules/transaction/category/category.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { CategoryController } from './category.controller';
-import { categoryProviders } from './category.provider';
-
-@Module({
-  controllers: [CategoryController],
-  providers: categoryProviders,
-})
-export class CategoryModule {}
