+++ src/metadata.ts
@@ -1,8 +1,8 @@
 /* eslint-disable */
 export default async () => {
     const t = {
-        ["./infra/http/nest/modules/transaction/category/dto/output.dto"]: await import("./infra/http/nest/modules/transaction/category/dto/output.dto"),
-        ["./core/modules/transaction/entities/category.entity"]: await import("./core/modules/transaction/entities/category.entity")
+        ["./modules/transaction/category/rest/dto/output.dto"]: await import("./modules/transaction/category/rest/dto/output.dto"),
+        ["./modules/transaction/category/core/entities/category.entity"]: await import("./modules/transaction/category/core/entities/category.entity")
     };
-    return { "@nestjs/swagger": { "models": [[import("./infra/http/nest/modules/transaction/category/dto/output.dto"), { "CategoryDto": { id: { required: true, type: () => String }, name: { required: true, type: () => String }, code: { required: true, type: () => Number }, description: { required: true, type: () => String }, type: { required: true, type: () => Object }, disabled: { required: true, type: () => Boolean }, blocked: { required: true, type: () => Boolean }, createdAt: { required: true, type: () => Date }, updatedAt: { required: true, type: () => Date } } }], [import("./infra/http/nest/common/interceptor/dto/response.dto"), { "CustomResponseDto": { success: { required: true, type: () => Boolean }, code: { required: true, type: () => Number }, data: { required: true, type: () => Object } } }], [import("./infra/http/nest/modules/transaction/category/dto/input.dto"), { "CreateCategoryDto": { name: { required: true, type: () => String }, description: { required: true, type: () => String }, type: { required: true, type: () => Object } } }]], "controllers": [[import("./infra/http/nest/modules/transaction/category/category.controller"), { "CategoryController": { "create": { type: t["./infra/http/nest/modules/transaction/category/dto/output.dto"].CategoryDto }, "findAll": { type: [t["./core/modules/transaction/entities/category.entity"].Category] }, "findOne": { type: t["./core/modules/transaction/entities/category.entity"].Category } } }]] } };
+    return { "@nestjs/swagger": { "models": [[import("./modules/transaction/category/rest/dto/input.dto"), { "CreateCategoryDto": { name: { required: true, type: () => String }, description: { required: true, type: () => String }, type: { required: true, type: () => Object } } }], [import("./modules/transaction/category/rest/dto/output.dto"), { "CategoryDto": { id: { required: true, type: () => String }, name: { required: true, type: () => String }, code: { required: true, type: () => Number }, description: { required: true, type: () => String }, type: { required: true, type: () => Object }, disabled: { required: true, type: () => Boolean }, blocked: { required: true, type: () => Boolean }, createdAt: { required: true, type: () => Date }, updatedAt: { required: true, type: () => Date } } }], [import("./common/interceptor/dto/response.dto"), { "CustomResponseDto": { success: { required: true, type: () => Boolean }, code: { required: true, type: () => Number }, data: { required: true, type: () => Object } } }]], "controllers": [[import("./modules/transaction/category/rest/category.controller"), { "CategoryController": { "create": { type: t["./modules/transaction/category/rest/dto/output.dto"].CategoryDto }, "findAll": { type: [t["./modules/transaction/category/core/entities/category.entity"].Category] }, "findOne": { type: t["./modules/transaction/category/core/entities/category.entity"].Category } } }]] } };
 };
