+++ src/modules/transaction/category/category.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { CategoryController } from './rest/category.controller';
+import { categoryProviders } from './category.provider';
+
+@Module({
+  controllers: [CategoryController],
+  providers: categoryProviders,
+})
+export class CategoryModule {}
