+++ src/modules/transaction/category/core/entities/category.entity.ts
