+++ src/modules/transaction/category/core/persistence/repository/base.interface.ts
