+++ src/modules/transaction/category/core/persistence/repository/category.interface.ts
