+++ src/modules/transaction/category/rest/dto/output.dto.ts
@@ -1,0 +1,47 @@
+import { CategoryType } from '@prisma/client';
+import {
+  IsBoolean,
+  IsDate,
+  IsIn,
+  IsNotEmpty,
+  IsNumber,
+  IsString,
+  IsUUID,
+} from 'class-validator';
+import { Category } from '../../core/entities/category.entity';
+
+export class CategoryDto {
+  @IsUUID('4')
+  id: string;
+
+  @IsString()
+  @IsNotEmpty()
+  name: string;
+
+  @IsNumber()
+  code: number;
+
+  @IsString()
+  @IsNotEmpty()
+  description: string;
+
+  @IsString()
+  @IsIn(['income', 'expense'])
+  type: CategoryType;
+
+  @IsBoolean()
+  disabled: boolean;
+
+  @IsBoolean()
+  blocked: boolean;
+
+  @IsDate()
+  createdAt: Date;
+
+  @IsDate()
+  updatedAt: Date;
+
+  constructor(category: Category) {
+    Object.assign(this, category.toJSON());
+  }
+}
