+++ tsconfig.json
@@ -28,8 +28,8 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["src/*"],
-      "@core/*": ["src/core/*"],
-      "@infra/*": ["src/infra/*"]
+      "@modules/*": ["src/modules/*"],
+      "@common/*": ["src/common/*"],
     },
     "types": ["vitest/globals"],
     "strict": true,
